test(vision): split the media-order cases out of msg_entry_tests.rs
They carried that file to 514 LoC, over the repo's 500 cap -- a NEW
violation, unlike the six the branch already carries. Moved to a sibling
`msg_entry_media_order_tests.rs` (the same #[path] idiom msg_entry_tests
itself uses), leaving 381 and 133.

No test changes: same 2220 pass, and the module path in the output moves
from msg_entry_tests::media_order_tests to msg_entry_media_order_tests.
